+++ test/babashka/bb_edn_test.clj
@@ -22,6 +22,14 @@
       :eof nil}
      ret)))
 
+(defn- run-exit [task & args]
+  (bb "-cp" "test-resources" "-e"
+      (format "(require '[babashka.cli :as cli])
+                (binding [cli/*exit-fn* (fn [m] (prn (select-keys m [:exit :cause])))
+                          *command-line-args* %s]
+                  (babashka.tasks/run '%s))"
+              (pr-str (vec args)) task)))
+
 (deftest deps-test
   (test-utils/with-config '{:deps {medley/medley {:mvn/version "1.3.0"}}}
     (is (= '{1 {:id 1}, 2 {:id 2}}
@@ -985,18 +993,10 @@ even more stuff here\"
                          "--count"))
       (testing ":require applies"
         (is (= {:exit 1 :cause :require}
-               (bb "-cp" "test-resources" "-e"
-                   "(require '[babashka.cli :as cli])
-                    (binding [cli/*exit-fn* (fn [m] (prn (select-keys m [:exit :cause])))
-                              *command-line-args* [\"deploy\"]]
-                      (babashka.tasks/run (quote db)))")))
+               (run-exit 'db "deploy")))
         (testing "in a nested :cmd"
           (is (= {:exit 1 :cause :require}
-                 (bb "-cp" "test-resources" "-e"
-                     "(require '[babashka.cli :as cli])
-                      (binding [cli/*exit-fn* (fn [m] (prn (select-keys m [:exit :cause])))
-                                *command-line-args* [\"helper\" \"local\"]]
-                        (babashka.tasks/run (quote db)))")))))
+                 (run-exit 'db "helper" "local")))))
       (testing "help prefers :doc in :org.babashka/cli over the docstring for a var and a symbol"
         (let [help (test-utils/bb nil "-cp" "test-resources" "db" "--help")]
           (is (not (str/includes? help "Get credentials")))
@@ -1021,11 +1021,7 @@ even more stuff here\"
                (bb "-cp" "test-resources" "go" "--port" "1" "--verbose"))))
       (testing "ns-level :restrict applies"
         (is (= {:exit 1 :cause :restrict}
-               (bb "-cp" "test-resources" "-e"
-                   "(require '[babashka.cli :as cli])
-                    (binding [cli/*exit-fn* (fn [m] (prn (select-keys m [:exit :cause])))
-                              *command-line-args* [\"--nope\"]]
-                      (babashka.tasks/run (quote go)))"))))))
+               (run-exit 'go "--nope"))))))
   (testing "a vector of [name command] pairs keeps the order it was written in"
     ;; more than 8 commands: an edn map would already have lost its order
     (test-utils/with-config '{:tasks {big {:cmd [["kilo" {:fn clojure.core/prn}]
@@ -1161,18 +1157,10 @@ even more stuff here\"
                (bb "-cp" "test-resources" "foo" "--port" "8080"))))
       (testing "an unknown option errors via the default :restrict"
         (is (= {:exit 1 :cause :restrict}
-               (bb "-cp" "test-resources" "-e"
-                   "(require '[babashka.cli :as cli])
-                    (binding [cli/*exit-fn* (fn [m] (prn (select-keys m [:exit :cause])))
-                              *command-line-args* [\"--nope\"]]
-                      (babashka.tasks/run (quote foo)))"))))
+               (run-exit 'foo "--nope"))))
       (testing "a stray positional errors via the default :restrict-args"
         (is (= {:exit 1 :cause :restrict-args}
-               (bb "-cp" "test-resources" "-e"
-                   "(require '[babashka.cli :as cli])
-                    (binding [cli/*exit-fn* (fn [m] (prn (select-keys m [:exit :cause])))
-                              *command-line-args* [\"prod\" \"extra\"]]
-                      (babashka.tasks/run (quote dep)))"))))))
+               (run-exit 'dep "prod" "extra"))))))
   (testing "a symbol :cli entry resolves to a defaults var, functions included"
     (test-utils/with-config '{:tasks {:cli babashka.tasks-cli/base-opts
                                       deploy {:cmd {"go" {:exec-fn babashka.tasks-cli/deploy-x}}}
@@ -1212,9 +1200,7 @@ even more stuff here\"
   (testing "dispatch errors reach a rebound *exit-fn*"
     (test-utils/with-config '{:tasks {deps {:cmd {"x" {:fn clojure.core/prn}}}}}
       (is (= {:exit 1 :cause :input-exhausted}
-             (bb "-e" "(require '[babashka.cli :as cli])
-                       (binding [cli/*exit-fn* (fn [m] (prn (select-keys m [:exit :cause])))]
-                         (babashka.tasks/run 'deps))")))))
+             (run-exit 'deps)))))
   (testing "a :cmd subcommand runs :depends, --help does not"
     (test-utils/with-config '{:tasks {dep {:task (println "DEP-RAN")}
                                       deps {:depends [dep]